Rename unithelper to unit_helper for consistency
