+++ changelog.md
@@ -3,6 +3,8 @@
 ## Unreleased
 - Fix: Add `naive_datetime()` test to the `pydantic-v2 branch`. (@bmeares, #522)
 - Fix: Add custom types that pass static type checks (@jsamoocha, #534)
+- Fix: Replace invariant with covariant container types (@lwasser, @jsamoocha, #510)
+- Fix: Rename unithelper.py --> unit_helper.py (@lwasser, #535)
 
 ## v1.7
