+++ src/stravalib/model.py
@@ -62,7 +62,7 @@ from stravalib.strava_model import (
     SportType,
     SummaryGear,
 )
-from stravalib.unithelper import _Quantity
+from stravalib.unit_helper import _Quantity
 
 if TYPE_CHECKING:
     from stravalib.client import BatchedResultsIterator
@@ -344,14 +344,14 @@ class _TimezoneAnnotation(_CustomStrAnnotation):
 
 
 class Distance(_Quantity):
-    """A class that supports conversion of units to meters. This output can then be 
-    converted to other units such as feet or meters using the stravalib.unit_helper module. 
-    
+    """A class that supports conversion of units to meters. This output can then be
+    converted to other units such as feet or meters using the stravalib.unit_helper module.
+
     Examples
     ----------
-    Once you have distance in meters, you can then use the unit_helper module to 
-    convert to other units such as feet 
-    
+    Once you have distance in meters, you can then use the unit_helper module to
+    convert to other units such as feet
+
     >>> from stravalib import unithelper
     >>> activity = client.get_activity(11416949675)
     >>> activity.distance.quantity()
@@ -361,6 +361,7 @@ class Distance(_Quantity):
     >>> unithelper.miles(activity.distance)
     <Quantity(5.00570419, 'mile')>
     """
+
     unit = "meters"
